PageInfo を routes 配下に吸収合併
ついでに QRCode 関係を lib に移動
