+++ src/lib/QRCode.svelte
