+++ src/lib/util/qrToSVG.ts
