+++ src/lib/util/qrToSVG.worker.ts
